Tutorials/12_handwriting_recognition_2/requirements.txt
tensorflow==2.19.0
tf-keras==2.19.0
tqdm==4.67.1
opencv-python==4.11.0.86
pillow==11.1.0
pandas==2.2.3
PyYAML==6.0.2
tf2onnx==1.16.1
onnxruntime==1.21.0
numpy==2.1.3